Alluser,cart page functionalities and jwt token creation complete

+++ bistro-boss-server/index.js
@@ -1,8 +1,10 @@
 const express = require('express')
 const admin = require("firebase-admin");
 const app = express()
+const jwt = require('jsonwebtoken')
 const cors = require('cors')
 require('dotenv').config()
+
 const port = process.env.PORT || 5000;
 
 //firebase admin
@@ -39,10 +41,18 @@ async function run() {
         const cartCollection = client.db("BistroBossDB").collection("carts")
 
         //api's
+
+        // jwt related api's 
+        app.post("/jwt", async (req, res) => {
+            const user = req.body;
+            const token = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET, { expiresIn: "24h" })
+            res.send({ token })
+        })
+
+        //users related api's
         app.post('/users', async (req, res) => {
             //insert email only if user doesn't exist
             const user = req.body;
-
             const query = { email: user.email }
             const existingUser = await usersCollection.findOne(query)
             if (existingUser) {
@@ -57,6 +67,18 @@ async function run() {
             res.send(result)
         })
 
+        app.patch("/users/admin/:id", async (req, res) => {
+            const id = req.params.id;
+            const filter = { _id: new ObjectId(id) };
+            const updatedDoc = {
+                $set: {
+                    role: 'admin'
+                }
+            }
+            const result = await usersCollection.updateOne(filter, updatedDoc)
+            res.send(result)
+        })
+
         app.delete('/users/:id', async (req, res) => {
             const id = req.params.id;
             const query = { _id: new ObjectId(id) };
@@ -65,6 +87,9 @@ async function run() {
             if (!user) {
                 return res.status(404).send({ message: "User not found" });
             }
+            if (!user.uid) {
+                return res.status(404).send({ message: "firebase uid not found" });
+            }
 
             //Delete the user from Firebase & mongodb
             await admin.auth().deleteUser(user?.uid);
@@ -72,14 +97,18 @@ async function run() {
             res.send(result)
         })
 
+        //menu related api's
         app.get('/menu', async (req, res) => {
             const result = await menuCollection.find().toArray()
             res.send(result)
         })
+        //review related api's
         app.get('/reviews', async (req, res) => {
             const result = await reviewsCollection.find().toArray()
             res.send(result)
         })
+
+        //cart related api's
         app.get('/carts', async (req, res) => {
             const email = req?.query?.email
             if (email) {
@@ -97,13 +126,13 @@ async function run() {
             const result = await cartCollection.insertOne(cartItem)
             res.send(result)
         })
-
         app.delete('/carts/:id', async (req, res) => {
             const id = req.params.id;
             const query = { _id: new ObjectId(id) }
             const result = await cartCollection.deleteOne(query)
             res.send(result)
         })
+
         // Send a ping to confirm a successful connection
         await client.db("admin").command({ ping: 1 });
         console.log("Pinged your deployment. You successfully connected to MongoDB!");
@@ -122,3 +151,5 @@ app.get('/', (req, res) => {
 app.listen(port, () => {
     console.log(`Bistro boss server is running on port ${port}`);
 })
+
+//require('crypto').randomBytes(64).toString('hex')
